+++ extension/src/time_vector/pipeline/delta.rs
@@ -52,7 +52,7 @@ pub fn timevector_delta<'s>(
             flags: series.flags,
             internal_padding: [0; 3],
             points: delta_points.into(),
-            null_val: std::vec::from_elem(0 as u8, nulls_len).into(),
+            null_val: std::vec::from_elem(0_u8, nulls_len).into(),
         }
     )
 }
